backend/requirements.txt
# Core dependencies
youtube-transcript-api==1.2.3
openai==2.8.1
faiss-cpu==1.7.4
numpy==1.24.3
boto3==1.34.51

# Text processing
tiktoken==0.6.0
langchain-text-splitters==0.0.1

# Utilities
python-dotenv==1.0.1
